requirements.txt
# Core ML/DL
torch>=2.0.0
transformers>=4.36.0
accelerate>=0.25.0
bitsandbytes>=0.41.0

# HuggingFace
huggingface-hub>=0.20.0
datasets>=2.16.0
tokenizers>=0.15.0

# Evaluation & Metrics
scikit-learn>=1.3.0
scipy>=1.11.0
numpy>=1.24.0

# Text Quality Metrics
nltk>=3.8.0
textstat>=0.7.3

# Data handling
pandas>=2.0.0
tqdm>=4.66.0

# Configuration & Environment
python-dotenv>=1.0.0
pyyaml>=6.0.0
hydra-core>=1.3.0

# Logging & Experiment tracking
wandb>=0.16.0
tensorboard>=2.15.0

# Visualization (optional)
matplotlib>=3.8.0
seaborn>=0.13.0

# For Binoculars detector
sentencepiece>=0.1.99

# Development
pytest>=7.4.0
black>=23.12.0
isort>=5.13.0
